Extended the PTrie Handling for Ads
