+++ data_logic/secret.py
@@ -1,0 +1,1 @@
+SECRET_KEY = "12345"
